--- .prettierrc
@@ -1,3 +1,0 @@
-{
-  "singleQuotes": true
-}
