+++ src/App.js
@@ -1,21 +1,20 @@
-/* eslint-disable */
-import React from "react";
-import { Routes, Route } from "react-router-dom";
-import Calculator from "./components/Calculator";
-import "./App.css";
-import Nav from "./components/Nav";
-import Home from "./components/Home";
-import Quote from "./components/Quote";
+import React from 'react';
+import { Routes, Route } from 'react-router-dom';
+import Calculator from './components/Calculator';
+import './App.css';
+import Nav from './components/Nav';
+import Home from './components/Home';
+import Quote from './components/Quote';
 
-class App extends React.Component {
+class App extends React.PureComponent {
   render() {
     return (
       <div className="App">
         <Nav />
         <Routes>
           <Route path="/home" element={<Home />} />
-          <Route path="calculator" element={<Calculator />} />
-          <Route path="about" element={<Quote />} />
+          <Route path="/calculator" element={<Calculator />} />
+          <Route path="/about" element={<Quote />} />
         </Routes>
       </div>
     );
